Add Skin Object add-on
This is a simple add-on for creating a single-edge object with a mirror, skin, and subsurf modifier added.

This makes it very quick and easy to start a sculpting basemesh.

The method for creating the single edge is very hacky, if anyone has a better way to do it I'd love to hear it. Currently it works by creating a plane, scaling that to 0 along the X axis, and then removing doubles. Surely there's a better way.

Signed-off-by: Jonathan Williamson <jonathan@cgcookie.com>