New Quick Mode Switch menu
This add-on gives a quick mode-switch menu with AL+TAB for changing between Object, Edit, and Sculpt modes.

This was originally a part of JayTools but I've stripped it out to its own add-on to make it easier to work with.

Currently the menus are static but I'm working on making them context sensitive.

Signed-off-by: Jonathan Williamson <jonathan@cgcookie.com>