ctrl + middle mouse to change number of spans
destination mesh behavior is as follows.  If in edit mode, new geometry will be added to current
object and user returned to edit mode.  Currently this requires that I toggle editmode, and use
bpy.ops.join to add in the geometry until the bmesh API catches up.  There are other ways I can
do it but for now I 'm going to stay consistent with bmesh and use this hack instead of back tracking
to bpy.data.mesh.  The only downside is that the selected objects may change.

If in object mode, a new object will be made and the
mesh data placed there.  User is returned to object mode so that the operator can be
undone or redone.

Banch Tag:  Alpha 1 release candidate for testing.