added a shift function in the space evenly on loop utility to allow fine grained shifting of the
simplified cross section.

Added ngon support by making a temporary object and triangulating Ngons.  Not the prettiest
solution but it does work.  Until obj.ray_cast returns the appropriate bmesh face index and not the
triangulated face index (which often doesn't exist in the bmesh) this is the most secure solution.