Fixes issue #61
Auto alignment can now be handled on an individual basis with the hot key 'A'

With the global auto alignment setting "off" the behavior is as follows.
If global auto aligment setting is on, this will still work, but will be overridden
if any new cuts are drawn.

'A' will align the hovered loop between the neighbors
'A' + 'Shift' will align the loop to the loop behind
'A' + 'Ctrl" will align the loop to the loop behind

The end loops will always be aligned to the loop behind or ahead

Cons:  User can not tell which direction is "ahead" or "behind" unless they remember which
cut they drew first. However, if they are attempting "diectional" alignment, the worst that can
happen is simply re-aligning with the other keyboard shortcut.