Closes issue #78
Improves issue #75.  Will leave issue open because this is a quick bandaid fix.
The underlying issue is the user's mesh is not a
water tight form in all directions, however the addon should be able to handle with grace those
situations and simply pass an error or warning without crashing and loosing all the previous
work.