Problem in logic image committed to documentation
Even spacing algorithm allows for subdivision as well as simplification
Slightly slower in the subdivision case.  Will optimize later or leave as is
unless complaints arise.

Fixes #90

This allows less dependence on the underlying base mesh
density which ultimately, is what we want the artist to not have
to worry about at all.